Tighten seams and team retry handling
