+++ src/mindroom/agents.py
@@ -21,6 +21,8 @@ from agno.session.team import TeamSession
 
 import mindroom.tools  # noqa: F401
 from mindroom import agent_prompts, constants
+from mindroom.agent_descriptions import describe_agent
+from mindroom.agent_storage import create_state_storage_db
 from mindroom.constants import ROUTER_AGENT_NAME
 from mindroom.credentials import get_runtime_credentials_manager
 from mindroom.hooks import HookRegistry
@@ -74,8 +76,6 @@ if TYPE_CHECKING:
 
 logger = get_logger(__name__)
 
-# Maximum length for instruction descriptions to include in agent summary
-_MAX_INSTRUCTION_LENGTH = 100
 _DEFAULT_MIND_AGENT_NAME = "mind"
 _DEFAULT_MIND_CONTEXT_FILES = (
     "SOUL.md",
@@ -666,14 +666,6 @@ def _build_dynamic_tooling_instruction_block(
     )
 
 
-def get_agent_runtime_sqlite_dbs(agent: Agent) -> tuple[SqliteDb | None, SqliteDb | None]:
-    """Return the runtime-owned SQLite handles attached to one agent."""
-    history_db = agent.db if isinstance(agent.db, SqliteDb) else None
-    learning = agent.learning
-    learning_db = learning.db if isinstance(learning, LearningMachine) and isinstance(learning.db, SqliteDb) else None
-    return history_db, learning_db
-
-
 def create_session_storage(
     agent_name: str,
     config: Config,
@@ -689,19 +681,6 @@ def create_session_storage(
         session_table=f"{agent_name}_sessions",
         execution_identity=execution_identity,
     )
-
-
-def create_state_storage_db(
-    storage_name: str,
-    state_root: Path,
-    *,
-    subdir: str,
-    session_table: str,
-) -> SqliteDb:
-    """Create a persistent SQLite database from an already-resolved state root."""
-    db_dir = state_root / subdir
-    db_dir.mkdir(parents=True, exist_ok=True)
-    return SqliteDb(session_table=session_table, db_file=str(db_dir / f"{storage_name}.db"))
 
 
 def enable_all_history_replay(entity: Agent | Team) -> None:
@@ -1274,67 +1253,6 @@ def create_agent(  # noqa: PLR0915, C901, PLR0912
     return agent
 
 
-def describe_agent(agent_name: str, config: Config) -> str:
-    """Generate a description of an agent or team based on its configuration.
-
-    Args:
-        agent_name: Name of the agent or team to describe
-        config: Application configuration
-
-    Returns:
-        Human-readable description of the agent or team
-
-    """
-    # Handle built-in router agent
-    if agent_name == ROUTER_AGENT_NAME:
-        return (
-            "router\n"
-            "  - Route messages to the most appropriate agent based on context and expertise.\n"
-            "  - Analyzes incoming messages and determines which agent is best suited to respond."
-        )
-
-    # Check if it's a team
-    if agent_name in config.teams:
-        team_config = config.teams[agent_name]
-        parts = [f"{agent_name}"]
-        if team_config.role:
-            parts.append(f"- {team_config.role}")
-        parts.append(f"- Team of agents: {', '.join(team_config.agents)}")
-        parts.append(f"- Collaboration mode: {team_config.mode}")
-        return "\n  ".join(parts)
-
-    # Check if agent exists
-    if agent_name not in config.agents:
-        return f"{agent_name}: Unknown agent or team"
-
-    agent_config = config.agents[agent_name]
-
-    # Start with agent name (not display name, for routing consistency)
-    parts = [f"{agent_name}"]
-    if agent_config.role:
-        parts.append(f"- {agent_config.role}")
-
-    # Add tools if any
-    effective_tools = config.get_agent_tools(agent_name)
-    if effective_tools:
-        tool_list = ", ".join(effective_tools)
-        parts.append(f"- Tools: {tool_list}")
-
-    # Add delegation targets if any
-    if agent_config.delegate_to:
-        delegate_list = ", ".join(agent_config.delegate_to)
-        parts.append(f"- Can delegate to: {delegate_list}")
-
-    # Add key instructions if any
-    if agent_config.instructions:
-        # Take first instruction as it's usually the most descriptive
-        first_instruction = agent_config.instructions[0]
-        if len(first_instruction) < _MAX_INSTRUCTION_LENGTH:  # Only include if reasonably short
-            parts.append(f"- {first_instruction}")
-
-    return "\n  ".join(parts)
-
-
 def get_agent_ids_for_room(
     room_key: str,
     config: Config,
@@ -1383,11 +1301,9 @@ __all__ = [
     "build_agent_toolkit",
     "create_agent",
     "create_session_storage",
-    "create_state_storage_db",
     "describe_agent",
     "ensure_default_agent_workspaces",
     "get_agent_ids_for_room",
-    "get_agent_runtime_sqlite_dbs",
     "get_agent_session",
     "get_agent_toolkit_names",
     "get_rooms_for_entity",
