+++ src/mindroom/custom_tools/delegate.py
@@ -13,7 +13,7 @@ from uuid import uuid4
 
 from agno.tools import Toolkit
 
-from mindroom.agents import describe_agent
+from mindroom.agent_descriptions import describe_agent
 from mindroom.ai import ai_response
 from mindroom.knowledge import KnowledgeManager, ensure_request_knowledge_managers, get_agent_knowledge
 from mindroom.logging_config import get_logger
