+++ src/mindroom/thread_summary.py
@@ -14,7 +14,8 @@ import nio
 from agno.agent import Agent
 from pydantic import BaseModel, Field
 
-from mindroom.ai import cached_agent_run, get_model_instance
+from mindroom.ai import get_model_instance
+from mindroom.ai_runtime import cached_agent_run
 from mindroom.logging_config import get_logger
 from mindroom.matrix.client_delivery import send_message_result
 from mindroom.matrix.message_builder import build_message_content
