+++ tests/test_agents.py
@@ -16,11 +16,11 @@ from agno.learn import LearningMachine, LearningMode, UserMemoryConfig, UserProf
 from pydantic import ValidationError
 
 from mindroom import agent_prompts
+from mindroom.agent_storage import get_agent_runtime_sqlite_dbs
 from mindroom.agents import (
     _CULTURE_MANAGER_CACHE,
     _PRIVATE_CULTURE_MANAGER_CACHE,
     create_agent,
-    get_agent_runtime_sqlite_dbs,
     get_agent_toolkit_names,
 )
 from mindroom.config.agent import (
