+++ tests/test_queued_message_notify.py
@@ -19,14 +19,12 @@ from agno.run.team import TeamRunOutput
 from agno.session.agent import AgentSession
 from agno.session.team import TeamSession
 
-from mindroom.ai import (
+from mindroom.ai import PreparedAgentRun, ai_response, stream_agent_response
+from mindroom.ai_runtime import (
     QUEUED_MESSAGE_NOTICE_TEXT,
-    PreparedAgentRun,
-    ai_response,
     cleanup_queued_notice_state,
     install_queued_message_notice_hook,
     queued_message_signal_context,
-    stream_agent_response,
 )
 from mindroom.bot import AgentBot
 from mindroom.bot_runtime_view import BotRuntimeState
